1.) rename MessageTreater to MessageHandler 2.) move MessageHandler to its own class 3.) added a couple comments about follow-up commits 4.) added two getters to facilitate the move to its own class
